Fix(Customer): Colocando regra de validação no UseCase e removendo do Repositório

+++ core/src/Application/UseCase/Customer/Delete.php
@@ -2,6 +2,7 @@
 
 namespace TechChallenge\Application\UseCase\Customer;
 
+use TechChallenge\Domain\Customer\Exceptions\CustomerNotFoundException;
 use TechChallenge\Domain\Customer\UseCase\Delete as ICustomerUseCaseDelete;
 use TechChallenge\Domain\Customer\Repository\ICustomer as ICustomerRepository;
 use TechChallenge\Domain\Customer\UseCase\DtoInput;
@@ -14,7 +15,10 @@ class Delete implements ICustomerUseCaseDelete
 
     public function execute(DtoInput $data): void
     {
-        $customer = $this->CustomerRepository->show($data->id);
+        if (!$this->CustomerRepository->exist(["id" => $data->id]))
+            throw new CustomerNotFoundException();
+
+        $customer = $this->CustomerRepository->show(["id" => $data->id]);
 
         $customer->delete();
